refactor: 지하철 역세권 지수 관련 오버레이 분리
